Back-End - Troca de banco, funções de cadastro e atualização de produtos, busca.
Troca de banco para o SQL SERVER;
Funções para listagem, inclusão e update de produtos;
Campos de busca por nome e categoria;
Código de conexão com banco de dados em arquivo separado;
Inclusão de id e tipo de usuário no session

+++ functions.php
@@ -1,103 +1,117 @@
 <?php	
 	session_start();
+	include('../db/index.php');
 	$errorMsg = '';
 
-	function newUser() {
-		if($db = odbc_connect("pidsn", "", "")) {
+	function newUser($db) {			
+		$password = checkFormInputs('password');
+		$login = checkFormInputs('email');
+		$name = checkFormInputs('name');
+
+		$query = odbc_exec($db, "SELECT loginUsuario FROM Usuario WHERE loginUsuario = '$login'");
+		$result = odbc_fetch_array($query);
+
+		if(!empty($result['loginUsuario'])) {
+			$GLOBALS['errorMsg'] = "Esse login já está cadastrado!";	
+		}			
+		else {
+			odbc_exec($db, "INSERT INTO usuario
+								(senha,
+								email,
+								nome)
+							VALUES 
+								('$password',
+								'$email',
+								'$name')");	
+			header("Location: ../login");
+		}
+	}
+
+	function detalhesItem($db) {        
+		$query = odbc_exec($db, "SELECT * FROM Produtos");
+        $result = odbc_fetch_array($query);
+        
+        $idCateg = $result['idCategoria'];
+        $subquery = odbc_exec($db, "SELECT * FROM Categoria WHERE idCategoria=$idCateg");
+        $subresult = odbc_fetch_array($subquery);
+            
+        echo "<table cellspacing='0' class=''>
+                <tr>
+                    <td class='bold'>Item</td>
+                    <td class='textocell'>", $result['nomeProduto'], "</td>
+                </tr>
+                <tr>
+                    <td class='bold'>Preço</td>
+                    <td class='textocell'>", $result['preco'], "</td>
+                </tr>
+                <tr>
+                    <td class='bold'>Categoria</td>
+                    <td class='textocell'>", $subresult['nameCategoria'], "</td>
+                </tr>
+                <tr>
+                    <td class='bold'>Estoque</td>
+                    <td class='textocell'>", $result['quantidade'], "</td>
+                </tr>
+                <tr>
+                    <td class='bold'>Descrição</td>
+                    <td class='textocell'>", $result['descricao'], "</td>
+                </tr>
+                <tr>
+                    <td class='bold'>Imagem</td>
+                    <td class='textocell'>", $result['imagem'], "</td>
+                </tr>";
+        
+    }
+    
+    
+    function searchItem($db) {		
+		$query = odbc_exec($db, "SELECT * FROM Produto");
+		echo "<table cellspacing='0' class=''>
+				<thead>
+					<th>Item</th>
+					<th>Preço</th>
+					<th>Categoria</th>
+                    <th>Ações</th>
+                </thead><tbody>";
+		while($result = odbc_fetch_array($query)) {
 			
-			$login = checkFormInputs('login');
-			$password = checkFormInputs('password');
-			$email = checkFormInputs('email');
-			$name = checkFormInputs('name');
-
-			$query = odbc_exec($db, "SELECT login FROM usuario WHERE login = '$login'");
-			$result = odbc_fetch_array($query);
-
-			if(!empty($result['login'])) {
-				$GLOBALS['errorMsg'] = "Esse login já está cadastrado!";	
-			}			
-			else {
-				odbc_exec($db, "INSERT INTO usuario
-									(login, 
-									senha,
-									email,
-									nome)
-								VALUES 
-									('$login',
-									'$password',
-									'$email',
-									'$name')");	
-				header("Location: ../login/index.php");
-			}
+           // $idCateg = $result['idCategoria'];
+            //$subquery = odbc_exec($db, "SELECT * FROM Categoria WHERE idCategoria=$idCateg");
+            //$subresult = odbc_fetch_array($subquery);
+            
+            echo "<tr>
+			        <td>", $result['nomeProduto'], "</td>
+			        <td>R$ ", $result['preco'], "</td>
+			        <td>", $subresult['nameCategoria'], "</td>
+                    <td id='acoes'>
+                        <a href='detalhes/?idItem=", $result['idProduto'],"'><div class='detalhes'><p>Detalhes</p></div></a>
+                        <a href='inserir/?idItem=", $result['idProduto'],"'><div class='edita'><p>Editar</p></div></a>
+                        <div class='deleta'><p>Excluir</p></div>
+                                        
+                    </td>
+			    </tr>";
 		}
-		else echo "Erro ao cadastrar usuário!";
-	}
-
-	function searchItem() {
-		if($db = odbc_connect("pidsn", "", "")) {
-			$query = odbc_exec($db, "SELECT * FROM usuario");
-			echo "<table cellspacing='0'>
-					<tr>
-						<th>Login</th>
-						<th>Senha</th>
-						<th>E-mail</th>
-						<th>Nome</th>
-					</tr>";
-			while($result = odbc_fetch_array($query)) {
-				echo "<tr>";
-					echo "<td>".$result['login']."</td>";
-					echo "<td>".$result['senha']."</td>";
-					echo "<td>".$result['email']."</td>";
-					echo "<td>".$result['nome']."</td>";
-				echo "</tr>";
-			}
-			echo "</table>";
-		}
-	}
-
-	function updateItem() {
-		if($db = odbc_connect("pidsn","","")) {
-			if(odbc_exec($db, "UPDATE Usuario
-						   SET
-						   loginUsuario = '1234@gmail.com',
-						   tipoPerfil = 'X'
-						   WHERE
-						   loginUsuario = 'khjkhjk@gmail.com'")) {
-			echo "Usuário atualizado com sucesso!<br>";
-							}
-		else echo "Erro ao atualizar usuário";		
-		}
-	}
-
-	function deleteItem() {
-		if($db = odbc_connect("pidsn","","")) {				
-			if(odbc_exec($db, "DELETE FROM Usuario
-							   WHERE
-							   loginUsuario = '1234@gmail.com'")) {
-				echo "Usuário deletado com sucesso!<br>";
-								}
-			else echo "Erro ao deletar usuário";			
-		}
-		else echo "Não Conectou!";
+		echo "</tbody> </table>";		
 	}
 	
-	function logIn() {
+	function logIn($db) {
 		if(!isset($_SESSION['user'])) {
 			if(isset($_POST['login'])) $login = checkFormInputs('login');
 			if(isset($_POST['password'])) $password = checkFormInputs('password');
-			if($db = odbc_connect("pidsn", "", "")) {
-				$query = odbc_exec($db, "SELECT nome FROM usuario WHERE login = '$login' AND senha = '$password'");
-				
-				$result = odbc_fetch_array($query);
+			
+			$query = odbc_exec($db, "SELECT nomeUsuario, idUsuario, tipoPerfil FROM Usuario WHERE loginUsuario = '$login' AND senhaUsuario = '$password'");
+			
+			$result = odbc_fetch_array($query);
 
-				if(!empty($result['nome'])) {
-					$_SESSION['user'] = $result['nome'];
-					header("Location: ../dashboard/index.php");
-				}
-				else $GLOBALS['errorMsg'] = "Email ou Senha Inválidos!";				
+			if(!empty($result['nomeUsuario'])) {
+				$_SESSION['user'] = $result['nomeUsuario'];
+				$_SESSION['id'] = $result['idUsuario']; // dizer quem cadastrou, chave estrangeira
+				$_SESSION['type'] = $result['tipoPerfil'];
+				header("Location: ../dashboard/");
 			}
+			else $GLOBALS['errorMsg'] = "Email ou Senha Inválidos!";			
 		}		
-	}
+    }
 
 	function logOut() {
 		if(isset($_GET['session'])) {
@@ -112,19 +126,29 @@
 	function getSessionUserName() {
 		return $_SESSION['user'];
 	}
-	function checkAction() {
-		$action = $_GET['action'];
-
-		switch($action) {
-			case "list":
-				searchItem();
-				break;
-			case "update":
-				updateItem();
-				break;
-			case "delete":
-				deleteItem();
-				break;			
-		}
+	function getSessionUserId() {
+		return $_SESSION['id'];
 	}
+	function getSessionUserType() {
+		return $_SESSION['type'];
+	}
+	function checkAction($db) {
+		
+        if(isset($_GET['action'])){
+            $action = $_GET['action'];
+
+            switch($action) {
+                case "list":
+                    searchItem($db);
+                    break;
+                case "update":
+                    updateItem();
+                    break;
+                case "delete":
+                    deleteItem();
+                    break;			
+            }
+        }
+	}
+
 ?>
